refactor: split model coverage cache module
