+++ crates/bitnet-cli/src/model_cache.rs
@@ -5,10 +5,10 @@ use bitnet_models::{
     capability_check::detect_capabilities,
     model_contracts::{BitnetModelContract, bitnet_model_contracts, find_bitnet_model_contract},
 };
-use clap::{Args, Subcommand, ValueEnum};
+use clap::{Args, Subcommand};
 use futures::StreamExt;
 use humansize::{DECIMAL, format_size};
-use serde::{Deserialize, Serialize};
+use serde::Serialize;
 use sha2::{Digest, Sha256};
 #[cfg(unix)]
 use std::process::Command;
@@ -19,9 +19,17 @@ use std::{
 };
 use tokio::io::AsyncWriteExt;
 
+mod coverage;
+
+pub use coverage::ModelStatusFormat;
+#[cfg(test)]
+use coverage::{
+    ModelStatusDashboard, ModelStatusRow, find_model_coverage_entry, model_status_dashboard,
+    read_model_coverage_matrix,
+};
+use coverage::{list_model_coverage, print_model_status};
+
 const DEFAULT_CACHE_RELATIVE: &[&str] = &["bitnet-rs", "models"];
-const MODEL_COVERAGE_MATRIX_RELATIVE: &[&str] =
-    &["ci", "model-artifacts", "model-coverage-matrix.toml"];
 const LOW_DISK_HEADROOM_BYTES: u64 = 1_073_741_824;
 #[cfg(feature = "full-cli")]
 pub(crate) const M4_SLM_RUNTIME_MODEL_ID: &str = "qwen2.5-0.5b-instruct-q8_0";
@@ -146,12 +154,6 @@ pub enum ModelAction {
         #[arg(long, default_value_t = false)]
         json: bool,
     },
-}
-
-#[derive(Clone, Copy, Debug, ValueEnum)]
-pub enum ModelStatusFormat {
-    Text,
-    Json,
 }
 
 #[derive(Debug, Clone, Copy, Serialize)]
@@ -437,119 +439,6 @@ struct PruneResult {
     existed: bool,
     removed: bool,
     dry_run: bool,
-}
-
-#[derive(Debug, Clone, Deserialize, Serialize)]
-struct ModelCoverageMatrix {
-    schema: u32,
-    artifact_kind: String,
-    updated: String,
-    work_item: String,
-    claim_boundary: String,
-    #[serde(default)]
-    tier: Vec<ModelCoverageTier>,
-    #[serde(default)]
-    entry: Vec<ModelCoverageEntry>,
-}
-
-#[derive(Debug, Clone, Deserialize, Serialize)]
-struct ModelCoverageTier {
-    id: String,
-    rank: u32,
-    #[serde(default)]
-    requires: Vec<String>,
-    meaning: String,
-}
-
-#[derive(Debug, Clone, Deserialize, Serialize)]
-struct ModelCoverageEntry {
-    id: String,
-    model_class: String,
-    family: String,
-    artifact_kind: String,
-    #[serde(default)]
-    contract_id: Option<String>,
-    #[serde(default)]
-    capability_id: Option<String>,
-    status: String,
-    current_tier: String,
-    verifier_surface: String,
-    tokenizer_authority: String,
-    prompt_authority: String,
-    cpu_reference: String,
-    #[serde(default)]
-    accelerator_routes: Vec<String>,
-    #[serde(default)]
-    required_receipts: Vec<String>,
-    #[serde(default)]
-    forbidden_claims: Vec<String>,
-    next_proof: String,
-    claim_boundary: String,
-    claims: ModelCoverageClaims,
-}
-
-#[derive(Debug, Clone, Deserialize, Serialize)]
-struct ModelCoverageClaims {
-    registered: bool,
-    structurally_valid: bool,
-    reference_good: bool,
-    cpu_answer_ready: bool,
-    accelerator_answer_ready: bool,
-    benchmark_qualified: bool,
-    product_cli_ready: bool,
-    server_ready: bool,
-    speedup_claim: bool,
-    full_residency_claim: bool,
-    bitnet_packed_i2s_qk256_proof: bool,
-    dense_regular_llm_cuda_proof: bool,
-}
-
-#[derive(Debug, Serialize)]
-struct ModelCoverageMatrixOutput<'a> {
-    matrix_path: &'a Path,
-    matrix: &'a ModelCoverageMatrix,
-}
-
-#[derive(Debug, Serialize)]
-struct ModelCoverageEntryOutput<'a> {
-    matrix_path: &'a Path,
-    entry: &'a ModelCoverageEntry,
-}
-
-#[derive(Debug, Serialize)]
-struct ModelStatusDashboard {
-    schema_version: u32,
-    device: String,
-    source: PathBuf,
-    note: &'static str,
-    models: Vec<ModelStatusRow>,
-}
-
-#[derive(Debug, Serialize)]
-struct ModelStatusRow {
-    id: String,
-    display_name: String,
-    model_class: String,
-    route: Option<String>,
-    tier: String,
-    status: String,
-    category: String,
-    cpu_answer_ready: bool,
-    accelerator_answer_ready: bool,
-    benchmark_qualified: bool,
-    speedup_claim: bool,
-    server_ready: bool,
-    full_residency_claim: bool,
-    bitnet_packed_i2s_qk256_proof: bool,
-    dense_regular_llm_cuda_proof: bool,
-    ask: String,
-    one_token: String,
-    short_decode: String,
-    warm_session: String,
-    benchmark: String,
-    server: String,
-    claim_boundary: String,
-    next_proof: String,
 }
 
 impl ModelCommand {
@@ -1093,381 +982,6 @@ fn print_contract_summary(summary: &VerifyContractSummary) {
     println!("permitted claims: {}", summary.permitted_claims.join(", "));
     println!("required receipts: {}", summary.required_receipts.join(", "));
     println!("claim boundary: {}", summary.claim_boundary);
-}
-
-fn list_model_coverage(id: Option<&str>, matrix: Option<PathBuf>, json: bool) -> Result<()> {
-    let matrix_path = resolve_model_coverage_matrix_path(matrix)?;
-    let matrix = read_model_coverage_matrix(&matrix_path)?;
-
-    if let Some(id) = id {
-        let entry = find_model_coverage_entry(&matrix, id).ok_or_else(|| {
-            let known = matrix.entry.iter().map(|entry| entry.id.as_str()).collect::<Vec<_>>();
-            anyhow!("unknown model coverage id `{id}`. Known coverage ids: {}", known.join(", "))
-        })?;
-        if json {
-            println!(
-                "{}",
-                serde_json::to_string_pretty(&ModelCoverageEntryOutput {
-                    matrix_path: &matrix_path,
-                    entry,
-                })?
-            );
-        } else {
-            print_model_coverage_entry(&matrix_path, entry);
-        }
-        return Ok(());
-    }
-
-    if json {
-        println!(
-            "{}",
-            serde_json::to_string_pretty(&ModelCoverageMatrixOutput {
-                matrix_path: &matrix_path,
-                matrix: &matrix,
-            })?
-        );
-        return Ok(());
-    }
-
-    print_model_coverage_overview(&matrix_path, &matrix);
-    Ok(())
-}
-
-fn resolve_model_coverage_matrix_path(matrix: Option<PathBuf>) -> Result<PathBuf> {
-    if let Some(path) = matrix {
-        return Ok(path);
-    }
-    if let Some(path) = std::env::var_os("BITNET_MODEL_COVERAGE_MATRIX") {
-        return Ok(PathBuf::from(path));
-    }
-    if let Ok(current_dir) = std::env::current_dir()
-        && let Some(path) = find_model_coverage_matrix_from(&current_dir)
-    {
-        return Ok(path);
-    }
-    if let Ok(exe) = std::env::current_exe()
-        && let Some(parent) = exe.parent()
-        && let Some(path) = find_model_coverage_matrix_from(parent)
-    {
-        return Ok(path);
-    }
-
-    anyhow::bail!(
-        "could not locate {}; run from the BitNet-rs repo, pass --matrix <PATH>, or set BITNET_MODEL_COVERAGE_MATRIX",
-        MODEL_COVERAGE_MATRIX_RELATIVE.join("/")
-    )
-}
-
-fn find_model_coverage_matrix_from(start: &Path) -> Option<PathBuf> {
-    for ancestor in start.ancestors() {
-        let mut candidate = ancestor.to_path_buf();
-        for segment in MODEL_COVERAGE_MATRIX_RELATIVE {
-            candidate.push(segment);
-        }
-        if candidate.exists() {
-            return Some(candidate);
-        }
-    }
-    None
-}
-
-fn read_model_coverage_matrix(path: &Path) -> Result<ModelCoverageMatrix> {
-    let text =
-        fs::read_to_string(path).with_context(|| format!("failed to read {}", path.display()))?;
-    let matrix: ModelCoverageMatrix = toml::from_str(&text)
-        .with_context(|| format!("failed to parse model coverage matrix {}", path.display()))?;
-    if matrix.schema != 1 {
-        anyhow::bail!("unsupported model coverage schema {}", matrix.schema);
-    }
-    if matrix.artifact_kind != "model_coverage_matrix" {
-        anyhow::bail!("expected artifact_kind=model_coverage_matrix, got {}", matrix.artifact_kind);
-    }
-    if matrix.tier.is_empty() {
-        anyhow::bail!("model coverage matrix has no tiers");
-    }
-    if matrix.entry.is_empty() {
-        anyhow::bail!("model coverage matrix has no entries");
-    }
-    Ok(matrix)
-}
-
-fn find_model_coverage_entry<'a>(
-    matrix: &'a ModelCoverageMatrix,
-    id: &str,
-) -> Option<&'a ModelCoverageEntry> {
-    matrix.entry.iter().find(|entry| entry.id.eq_ignore_ascii_case(id))
-}
-
-fn print_model_coverage_overview(path: &Path, matrix: &ModelCoverageMatrix) {
-    println!("Model coverage matrix: {} entries, {} tiers", matrix.entry.len(), matrix.tier.len());
-    println!("source: {}", path.display());
-    println!("updated: {} ({})", matrix.updated, matrix.work_item);
-    println!("claim boundary: {}", matrix.claim_boundary);
-    println!();
-    println!("{:<42} {:<20} {:<20} {:<28} Routes", "ID", "Class", "Tier", "Status");
-    println!("{}", "-".repeat(128));
-    for entry in &matrix.entry {
-        let routes = if entry.accelerator_routes.is_empty() {
-            "-".to_string()
-        } else {
-            entry.accelerator_routes.join(", ")
-        };
-        println!(
-            "{:<42} {:<20} {:<20} {:<28} {}",
-            entry.id, entry.model_class, entry.current_tier, entry.status, routes
-        );
-    }
-    println!();
-    println!("Use `bitnet model coverage <id>` for one row, or add --json for receipts tooling.");
-}
-
-fn print_model_coverage_entry(path: &Path, entry: &ModelCoverageEntry) {
-    println!("coverage: {}", entry.id);
-    println!("source: {}", path.display());
-    println!("class: {} / {}", entry.model_class, entry.family);
-    println!("artifact: {}", entry.artifact_kind);
-    if let Some(contract_id) = &entry.contract_id {
-        println!("contract: {contract_id}");
-    }
-    if let Some(capability_id) = &entry.capability_id {
-        println!("capability: {capability_id}");
-    }
-    println!("status: {}", entry.status);
-    println!("tier: {}", entry.current_tier);
-    println!("verifier: {}", entry.verifier_surface);
-    println!("tokenizer authority: {}", entry.tokenizer_authority);
-    println!("prompt authority: {}", entry.prompt_authority);
-    println!("cpu reference: {}", entry.cpu_reference);
-    if entry.accelerator_routes.is_empty() {
-        println!("routes: -");
-    } else {
-        println!("routes: {}", entry.accelerator_routes.join(", "));
-    }
-    println!("required receipts: {}", entry.required_receipts.join(", "));
-    println!("forbidden claims: {}", entry.forbidden_claims.join(", "));
-    println!("next proof: {}", entry.next_proof);
-    println!("claim boundary: {}", entry.claim_boundary);
-    println!("claims:");
-    println!("  registered: {}", entry.claims.registered);
-    println!("  structurally_valid: {}", entry.claims.structurally_valid);
-    println!("  reference_good: {}", entry.claims.reference_good);
-    println!("  cpu_answer_ready: {}", entry.claims.cpu_answer_ready);
-    println!("  accelerator_answer_ready: {}", entry.claims.accelerator_answer_ready);
-    println!("  benchmark_qualified: {}", entry.claims.benchmark_qualified);
-    println!("  product_cli_ready: {}", entry.claims.product_cli_ready);
-    println!("  server_ready: {}", entry.claims.server_ready);
-    println!("  speedup_claim: {}", entry.claims.speedup_claim);
-    println!("  full_residency_claim: {}", entry.claims.full_residency_claim);
-    println!("  bitnet_packed_i2s_qk256_proof: {}", entry.claims.bitnet_packed_i2s_qk256_proof);
-    println!("  dense_regular_llm_cuda_proof: {}", entry.claims.dense_regular_llm_cuda_proof);
-}
-
-fn print_model_status(
-    device: &str,
-    matrix: Option<PathBuf>,
-    format: ModelStatusFormat,
-) -> Result<()> {
-    let matrix_path = resolve_model_coverage_matrix_path(matrix)?;
-    let matrix = read_model_coverage_matrix(&matrix_path)?;
-    let dashboard = model_status_dashboard(device, &matrix_path, &matrix);
-
-    match format {
-        ModelStatusFormat::Json => {
-            println!("{}", serde_json::to_string_pretty(&dashboard)?);
-        }
-        ModelStatusFormat::Text => print_model_status_text(&dashboard),
-    }
-
-    Ok(())
-}
-
-fn model_status_dashboard(
-    device: &str,
-    matrix_path: &Path,
-    matrix: &ModelCoverageMatrix,
-) -> ModelStatusDashboard {
-    let models = matrix
-        .entry
-        .iter()
-        .filter(|entry| model_status_includes_entry(device, entry))
-        .map(model_status_row)
-        .collect();
-
-    ModelStatusDashboard {
-        schema_version: 1,
-        device: device.to_string(),
-        source: matrix_path.to_path_buf(),
-        note: "Read-only model coverage view; it does not probe hardware or create new proof.",
-        models,
-    }
-}
-
-fn model_status_includes_entry(device: &str, entry: &ModelCoverageEntry) -> bool {
-    if device != "nvidia-rtx-5070-ti-cuda" {
-        return false;
-    }
-    if entry.claims.product_cli_ready
-        && (entry.accelerator_routes.iter().any(|route| route == "bitnet_qk256_cuda")
-            || entry.accelerator_routes.iter().any(|route| route == "dense_regular_llm_cuda"))
-    {
-        return true;
-    }
-
-    !entry.claims.product_cli_ready
-        && entry.status.contains("candidate")
-        && matches!(entry.model_class.as_str(), "dense_slm" | "small_llm")
-}
-
-fn model_status_row(entry: &ModelCoverageEntry) -> ModelStatusRow {
-    let route = entry.accelerator_routes.first().cloned();
-    let category =
-        if entry.claims.product_cli_ready { "supported" } else { "candidate" }.to_string();
-    let benchmark = benchmark_status(entry);
-    let warm_session = warm_session_status(entry);
-    let ask = ask_status(entry);
-    let one_token = dense_receipt_status(entry, "one_token");
-    let short_decode = dense_receipt_status(entry, "short_decode");
-
-    ModelStatusRow {
-        id: entry.id.clone(),
-        display_name: model_status_display_name(entry),
-        model_class: entry.model_class.clone(),
-        route,
-        tier: entry.current_tier.clone(),
-        status: entry.status.clone(),
-        category,
-        cpu_answer_ready: entry.claims.cpu_answer_ready,
-        accelerator_answer_ready: entry.claims.accelerator_answer_ready,
-        benchmark_qualified: entry.claims.benchmark_qualified,
-        speedup_claim: entry.claims.speedup_claim,
-        server_ready: entry.claims.server_ready,
-        full_residency_claim: entry.claims.full_residency_claim,
-        bitnet_packed_i2s_qk256_proof: entry.claims.bitnet_packed_i2s_qk256_proof,
-        dense_regular_llm_cuda_proof: entry.claims.dense_regular_llm_cuda_proof,
-        ask,
-        one_token,
-        short_decode,
-        warm_session,
-        benchmark,
-        server: if entry.claims.server_ready { "ready" } else { "not ready" }.to_string(),
-        claim_boundary: entry.claim_boundary.clone(),
-        next_proof: entry.next_proof.clone(),
-    }
-}
-
-fn print_model_status_text(dashboard: &ModelStatusDashboard) {
-    println!("CUDA model status for {}", dashboard.device);
-    println!("source: {}", dashboard.source.display());
-    println!("{}", dashboard.note);
-    println!();
-
-    print_model_status_group(dashboard, "Supported", "supported");
-    println!();
-    print_model_status_group(dashboard, "Candidates", "candidate");
-}
-
-fn print_model_status_group(dashboard: &ModelStatusDashboard, title: &str, category: &str) {
-    println!("{title}:");
-    let mut printed = false;
-    for row in dashboard.models.iter().filter(|row| row.category == category) {
-        printed = true;
-        println!("  {}", row.display_name);
-        println!("    id: {}", row.id);
-        println!("    class: {}", model_status_class_label(&row.model_class));
-        println!("    route: {}", row.route.as_deref().unwrap_or("not ready"));
-        println!("    tier: {}", row.tier);
-        println!("    cpu answer: {}", ready_label(row.cpu_answer_ready));
-        println!("    cuda answer: {}", ready_label(row.accelerator_answer_ready));
-        if row.model_class == "dense_slm" && row.category == "supported" {
-            println!("    one-token: {}", row.one_token);
-            println!("    short-decode: {}", row.short_decode);
-        } else {
-            println!("    ask: {}", row.ask);
-        }
-        println!("    warm-session: {}", row.warm_session);
-        println!("    benchmark: {}", row.benchmark);
-        println!("    speedup: {}", if row.speedup_claim { "qualified" } else { "not qualified" });
-        println!("    server: {}", row.server);
-        println!(
-            "    full residency: {}",
-            if row.full_residency_claim { "claimed" } else { "not claimed" }
-        );
-        println!("    claim boundary: {}", row.claim_boundary);
-        if row.category == "candidate" {
-            println!("    next proof: {}", row.next_proof);
-        }
-        println!();
-    }
-
-    if !printed {
-        println!("  none");
-    }
-}
-
-fn model_status_display_name(entry: &ModelCoverageEntry) -> String {
-    if let Some(id) = &entry.capability_id {
-        return id.clone();
-    }
-    if let Some(id) = entry.verifier_surface.split_whitespace().last()
-        && !id.is_empty()
-        && id != "only"
-        && id != "matrix"
-    {
-        return id.to_string();
-    }
-    entry.contract_id.clone().unwrap_or_else(|| entry.id.clone())
-}
-
-fn model_status_class_label(model_class: &str) -> &'static str {
-    match model_class {
-        "bitnet" => "BitNet",
-        "dense_slm" => "dense SLM",
-        "small_llm" => "small dense LLM",
-        "modern_llm_docs_only" => "docs-only modern LLM",
-        _ => "model",
-    }
-}
-
-fn ready_label(ready: bool) -> &'static str {
-    if ready { "ready" } else { "not ready" }
-}
-
-fn ask_status(entry: &ModelCoverageEntry) -> String {
-    if entry.claims.product_cli_ready && entry.claims.accelerator_answer_ready {
-        "ready".to_string()
-    } else {
-        "not ready".to_string()
-    }
-}
-
-fn dense_receipt_status(entry: &ModelCoverageEntry, receipt_fragment: &str) -> String {
-    if entry.required_receipts.iter().any(|receipt| receipt.contains(receipt_fragment)) {
-        "ready".to_string()
-    } else {
-        "not ready".to_string()
-    }
-}
-
-fn warm_session_status(entry: &ModelCoverageEntry) -> String {
-    if entry.required_receipts.iter().any(|receipt| receipt.contains("warm_session"))
-        && entry.claims.accelerator_answer_ready
-    {
-        "ready".to_string()
-    } else {
-        "not ready".to_string()
-    }
-}
-
-fn benchmark_status(entry: &ModelCoverageEntry) -> String {
-    if entry.claims.benchmark_qualified && entry.claims.speedup_claim {
-        return "qualified".to_string();
-    }
-    if entry.claims.product_cli_ready
-        && entry.required_receipts.iter().any(|receipt| receipt.contains("benchmark"))
-    {
-        return "reviewed, speedup not accepted".to_string();
-    }
-    "not ready".to_string()
 }
 
 fn list_models(cache_dir: Option<PathBuf>, json: bool) -> Result<()> {
